requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.21.0
openai>=1.0.0 